refactor: move xclim indicator scanning into openeo/xclim_processes.py
