+++ open_climate_service/openeo/plugin_processes.py
@@ -11,7 +11,8 @@ from types import ModuleType
 from typing import Any
 
 from open_climate_service import config as api_config
-from open_climate_service.process import _OCS_PROCESS_ATTR, get_process_metadata
+from open_climate_service.openeo import xclim_processes
+from open_climate_service.process import get_process_metadata
 
 logger = logging.getLogger(__name__)
 
@@ -25,7 +26,7 @@ def load_plugin_processes() -> list[tuple[str, Any]]:
     3. Instance plugins — ``plugins_dir/processes/`` (override everything)
     """
     found: dict[str, Any] = {}
-    for func in _scan_xclim_indicators():
+    for func in xclim_processes.scan():
         meta = get_process_metadata(func)
         if meta:
             found[meta["id"]] = func
@@ -90,98 +91,3 @@ def _load_from_module(module_name: str) -> list[Any]:
     except Exception:
         logger.warning("Failed to load plugin processes from %s", module_name, exc_info=True)
         return []
-
-
-_xclim_indicator_cache: list[Any] | None = None
-
-
-def _scan_xclim_indicators() -> list[Any]:
-    """Auto-register all xclim indicators as process callables (result is cached)."""
-    global _xclim_indicator_cache
-    if _xclim_indicator_cache is not None:
-        return _xclim_indicator_cache
-    try:
-        _xclim_indicator_cache = _collect_xclim_indicators()
-    except Exception:
-        logger.warning("Failed to load xclim indicators", exc_info=True)
-        _xclim_indicator_cache = []
-    return _xclim_indicator_cache
-
-
-def _collect_xclim_indicators() -> list[Any]:
-    import inspect
-
-    import xclim.indicators.atmos as atmos
-    import xclim.indicators.land as land
-    import xclim.indicators.seaIce as seaice
-    from xclim.core.indicator import InputKind
-
-    _SKIP_KINDS = {InputKind.DATASET, InputKind.KWARGS}
-    # Kinds whose default is a dataset variable name, not a real scalar default
-    _VARIABLE_KINDS = {InputKind.VARIABLE, InputKind.OPTIONAL_VARIABLE}
-    _KIND_SCHEMA: dict[InputKind, dict[str, str]] = {
-        InputKind.VARIABLE: {"type": "object", "subtype": "datacube"},
-        InputKind.OPTIONAL_VARIABLE: {"type": "object", "subtype": "datacube"},
-        InputKind.QUANTIFIED: {"type": "string"},
-        InputKind.FREQ_STR: {"type": "string"},
-        InputKind.NUMBER: {"type": "number"},
-        InputKind.STRING: {"type": "string"},
-        InputKind.DAY_OF_YEAR: {"type": "string"},
-        InputKind.DATE: {"type": "string"},
-        InputKind.NUMBER_SEQUENCE: {"type": "array"},
-        InputKind.BOOL: {"type": "boolean"},
-    }
-
-    funcs: list[Any] = []
-    seen: set[str] = set()
-
-    for mod in (atmos, land, seaice):
-        for obj in vars(mod).values():
-            if not (hasattr(obj, "identifier") and hasattr(obj, "parameters") and hasattr(obj, "title")):
-                continue
-            indicator_id: str = obj.identifier
-            if indicator_id in seen:
-                continue
-            seen.add(indicator_id)
-
-            params: list[dict[str, Any]] = []
-            for name, p in obj.parameters.items():
-                if p.kind in _SKIP_KINDS:
-                    continue
-                param_meta: dict[str, Any] = {"name": name, "schema": _KIND_SCHEMA.get(p.kind, {})}
-                if p.description:
-                    param_meta["description"] = p.description
-                # Only expose defaults that are real JSON-serializable scalars.
-                # VARIABLE/OPTIONAL_VARIABLE defaults are dataset-mode variable name
-                # strings (e.g. "pr"), not meaningful API defaults.
-                if p.kind not in _VARIABLE_KINDS:
-                    default = p.default
-                    if default is not inspect.Parameter.empty and isinstance(
-                        default, (str, int, float, bool, type(None))
-                    ):
-                        param_meta["optional"] = True
-                        param_meta["default"] = default
-                elif p.kind is InputKind.OPTIONAL_VARIABLE:
-                    param_meta["optional"] = True
-                params.append(param_meta)
-
-            meta: dict[str, Any] = {
-                "id": indicator_id,
-                "summary": obj.title,
-                "description": obj.abstract,
-                "parameters": params,
-                "returns": {"schema": {}},
-            }
-            funcs.append(_make_indicator_callable(obj, meta))
-
-    return funcs
-
-
-def _make_indicator_callable(indicator: Any, meta: dict[str, Any]) -> Any:
-    """Wrap an xclim indicator as a bare callable with process metadata attached."""
-
-    def _call(**kwargs: Any) -> Any:
-        return indicator(**kwargs)
-
-    setattr(_call, _OCS_PROCESS_ATTR, meta)
-    return _call
